requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1
pytest-asyncio==0.22.0
playwright==1.44.0
pytest-playwright==0.5.2
sqlalchemy==2.0.23
alembic==1.12.1
passlib[bcrypt]==1.7.4
email-validator==2.3.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
